[MAINT] Move routing facilities into dedicated package
